+++ mmdet/models/detectors/base.py
@@ -200,7 +200,7 @@ class BaseDetector(BaseModule, metaclass=ABCMeta):
 
         # If the loss_vars has different length, GPUs will wait infinitely
         if dist.is_available() and dist.is_initialized():
-            log_var_length = torch.tensor(len(log_vars), device=loss.device).int()
+            log_var_length = torch.tensor(len(log_vars), device=loss.device)
             dist.all_reduce(log_var_length)
             message = (f'rank {dist.get_rank()}' +
                        f' len(log_vars): {len(log_vars)}' + ' keys: ' +
